requirements.txt
mysql-connector-python
customtkinter
pillow
tkcalendar
